+++ src/main/java/insane96mcp/runeenchanting/setup/RERunes.java
@@ -54,6 +54,7 @@ public class RERunes {
     public static final DeferredHolder<Rune, SwiftStrikeRune> SWIFT_STRIKE = register("swift_strike", new SwiftStrikeRune());
     public static final DeferredHolder<Rune, AirStealerRune> AIR_STEALER = register("air_stealer", new AirStealerRune());
     public static final DeferredHolder<Rune, ExplosiveRune> EXPLOSIVE = register("explosive", new ExplosiveRune());
+    public static final DeferredHolder<Rune, ElectrocutionRune> ELECTROCUTION = register("electrocution", new ElectrocutionRune());
     public static final DeferredHolder<Rune, PartBreakerRune> PART_BREAKER = register("part_breaker", new PartBreakerRune());
     public static final DeferredHolder<Rune, PaddingRune> PADDING = register("padding", new PaddingRune());